+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+# [1.219.0](https://github.com/Basis-Theory/developers.basistheory.com/compare/v1.218.0...v1.219.0) (2025-03-07)
+
+
+### Features
+
+* **web-elements:** improve example when loading via cdn ([#515](https://github.com/Basis-Theory/developers.basistheory.com/issues/515)) ([a014a65](https://github.com/Basis-Theory/developers.basistheory.com/commit/a014a65fc6f62afc4113b4ee9a6d1ea01b4f7e25))
+
+# [1.218.0](https://github.com/Basis-Theory/developers.basistheory.com/compare/v1.217.0...v1.218.0) (2025-03-07)
+
+
+### Features
+
+* add react-elements ([#514](https://github.com/Basis-Theory/developers.basistheory.com/issues/514)) ([7cdab8d](https://github.com/Basis-Theory/developers.basistheory.com/commit/7cdab8d75bf97950dc6334134b16be7342e517ba))
+
+# [1.217.0](https://github.com/Basis-Theory/developers.basistheory.com/compare/v1.216.0...v1.217.0) (2025-03-06)
+
+
+### Features
+
+* adding docs for get token intent ([#516](https://github.com/Basis-Theory/developers.basistheory.com/issues/516)) ([f7bd5a1](https://github.com/Basis-Theory/developers.basistheory.com/commit/f7bd5a14783e73963571011ca2c4c131aa4323d5))
+
 # [1.216.0](https://github.com/Basis-Theory/developers.basistheory.com/compare/v1.215.1...v1.216.0) (2025-02-20)